correctly map error pointers to the original document
when serializing ActiveModel::Errors, wrap it in an internal class to
associate the corresponding reverse mapping to that object so that the
serializer has access to it.

this does raise some questions with regards to configuring the
serializing class for ActiveModel::Errors. There is hidden knowledge
here that in order to change the serializing class that it must be
configured for the new wrapping JSONAPI::Rails::ActiveModel::Errors
class. Perhaps we should only use this wrapping class when the
configuration matches the defaults?

TODO: fix rendering for hashes (which we forgot to address, right now
ALL class types are wrapped as if they were ActiveModel::Errors) - the
tests are broken to remind us of this