Implement support for bulk extension (#1)
* Add bulk to supported extensions in header

* copy + convert deserializable_resource to handle multiple resources

* update reverse mapping pointers to support multiple resources

adding an index so that errors can be traced to their original source.
This follows the format defined in RFC6901
https://tools.ietf.org/html/rfc6901#page-5

* Escape bulk reverse mapping

* Merge deserializable_resources into deserializable_resource, add bulk check

* Don't add ext support in register_renderers

* use configuration to drive branching

* refactor detection of requested extensions

* More tests, small fix to how Content-Type string is concatenated

* Typo fixes

* Add more tests related to bulk extension

* Fix regex, add errors tests

* Comment out error test, fix extension_request logic

* A bunch of little things

* correctly map error pointers to the original document

when serializing ActiveModel::Errors, wrap it in an internal class to
associate the corresponding reverse mapping to that object so that the
serializer has access to it.

this does raise some questions with regards to configuring the
serializing class for ActiveModel::Errors. There is hidden knowledge
here that in order to change the serializing class that it must be
configured for the new wrapping JSONAPI::Rails::ActiveModel::Errors
class. Perhaps we should only use this wrapping class when the
configuration matches the defaults?

TODO: fix rendering for hashes (which we forgot to address, right now
ALL class types are wrapped as if they were ActiveModel::Errors) - the
tests are broken to remind us of this

* Deal with Hash errors, fix a few typos

* tweaks

* Account for empty Content-Type

* Updates after CR comments

* Add invariant violation, update test, add comments

* only restrict render :jsonai_errors lenght if bulk response

* Errors must be a flattened array

* add tests for mixed and sparse error rendering

* Move error formatting into its own method

Co-authored-by: Julien Fitzpatrick <jfitzpatrick@artemisag.com>